package/genimage: use git version, bump to latest
Using the latest release 7 of genimage on Arch Linux at the
moment seems to result in segmentation fault, while it works on
other distros. The latest git version on the other hand runs
correctly on all platforms tested. Test was done by running:
"make wandboard_defconfig && make" as Wandboard is using genimage
in the post-image script.